backend/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
